All flow control for Types of food
I couldn't sleep, so i made some flow control for Types of food, with all of them getting done, with arbitrary quantity amounts, and 3 items per section, with pictures, and the pages updating depending if more are needed, if current grid is filled. Location was also done, shows on google maps where its located. Need to do FAQ and small side buttons that are the same as current essentials website. Lastly need to make an edit or add item page.