API Updates and Backend Demo
Added Ash's menu with some small changes to make endpoints match. Updated more of the API. Changed the endpoints from Item for popular items and categories to CheckoutOrder and CheckoutItem. Refactored a bunch of code to ensure better responsibility handling and decouple things more, as well as ensure better presentation to the view. Going to refactor PurchaseOrder and PurchaseItem, may adjust some others as I clean up a bit more.